Added capability to edit different xml schema files (refs #151)

+++ www/devel/schema/oaSchema.php
@@ -81,26 +81,46 @@ class Openads_Schema_Manager
     /**
      * php5 class constructor
      *
+     * @param string The XML schema file we are working on
      */
-    function __construct()
+    function __construct($schema_file = 'core')
     {
-        $this->file_schema_core = 'tables_core.xml';
         $this->path_schema_final = MAX_PATH.'/etc/';
         $this->path_schema_trans = MAX_PATH.'/var/';
-
-        $this->file_changes_core = 'changes_core.xml';
         $this->path_changes_final = MAX_PATH.'/etc/changes/';
         $this->path_changes_trans = MAX_PATH.'/var/';
 
-        $this->file_links_core = 'db_schema.links.ini';
         $this->path_links_final = MAX_PATH.'/lib/max/Dal/DataObjects/';
         $this->path_links_trans = MAX_PATH.'/var/';
+
+        $this->xml_files = array();
+        foreach (glob($this->path_schema_final.'tables_*.xml') as $xml_file) {
+            $xml_file = basename($xml_file);
+            if (preg_match('/^tables_([^.]+)\.xml$/', $xml_file, $matches)) {
+                $this->xml_files[] = $matches[1];
+            }
+        }
+
+        if (in_array($schema_file, $this->xml_files)) {
+            $this->file_schema_core  = 'tables_'.$schema_file.'.xml';
+            $this->file_changes_core = 'changes_'.$schema_file.'.xml';
+        } else {
+            die ("Unknown xml file");
+        }
+
+        if ($schema_file == 'core') {
+            $this->file_links_core = 'db_schema.links.ini';
+        } else {
+            $this->file_links_core = '';
+        }
 
         $this->schema_final = $this->path_schema_final.$this->file_schema_core;
         $this->schema_trans = $this->path_schema_trans.$this->file_schema_core;
 
-        $this->links_final = $this->path_links_final.$this->file_links_core;
-        $this->links_trans = $this->path_links_trans.$this->file_links_core;
+        if (!empty($this->file_links_core)) {
+            $this->links_final = $this->path_links_final.$this->file_links_core;
+            $this->links_trans = $this->path_links_trans.$this->file_links_core;
+        }
 
         $this->changes_final = $this->path_changes_final.$this->file_changes_core;
         $this->changes_trans = $this->path_changes_trans.$this->file_changes_core;
@@ -122,11 +142,11 @@ class Openads_Schema_Manager
     /**
      * php4 class constructor
      *
-     * @return Openads_Schema_Manager
+     * @param string The XML schema file we are working on
      */
-    function Openads_Schema_Manager()
+    function Openads_Schema_Manager($schema_file = 'core')
     {
-        $this->__construct();
+        $this->__construct($schema_file);
     }
 
     /**
@@ -173,7 +193,7 @@ class Openads_Schema_Manager
                 $this->version++;
                 $this->writeWorkingDefinitionFile();
             }
-            if (file_exists($this->links_final))
+            if (!empty($this->links_final) && file_exists($this->links_final))
             {
                 copy($this->links_final, $this->links_trans);
             }
@@ -192,7 +212,7 @@ class Openads_Schema_Manager
     function commitFinal()
     {
         if ( (file_exists($this->schema_trans)) &&
-             (file_exists($this->links_trans))
+             (empty($this->links_trans) || file_exists($this->links_trans))
            )
         {
             $this->setWorkingFiles();
@@ -209,7 +229,7 @@ class Openads_Schema_Manager
             }
             if ($valid)
             {
-                if (file_exists($this->links_trans))
+                if (!empty($this->links_trans) && file_exists($this->links_trans))
                 {
                     copy($this->links_trans, $this->links_final);
                 }
@@ -237,13 +257,13 @@ class Openads_Schema_Manager
             $this->working_file_schema = $this->schema_final;
         }
 
-        if (file_exists($this->links_trans))
+        if (!empty($this->links_trans) && file_exists($this->links_trans))
         {
             $this->working_file_links = $this->links_trans;
         }
-        else if (file_exists($this->links_final))
+        else if (!empty($this->links_final) && file_exists($this->links_final))
         {
-            $this->working_file_links = $working_links_final;
+            $this->working_file_links = $this->links_final;
         }
         return (file_exists($this->working_file_links) && file_exists($this->working_file_schema));
     }
@@ -308,7 +328,7 @@ class Openads_Schema_Manager
      */
     function deleteLinksTrans()
     {
-        if (file_exists($this->links_trans))
+        if (!empty($this->links_final) && file_exists($this->links_trans))
         {
             unlink($this->links_trans);
         }
@@ -626,9 +646,13 @@ class Openads_Schema_Manager
      */
     function linkDelete($table_name, $link_name)
     {
-        $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
-        unset($links[$table_name][$link_name]);
-        return Openads_Links::writeLinksDotIni($this->links_trans, $links);
+        if (!empty($this->links_trans)) {
+            $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
+            unset($links[$table_name][$link_name]);
+            return Openads_Links::writeLinksDotIni($this->links_trans, $links);
+
+        }
+        return true;
     }
 
     /**
@@ -641,9 +665,12 @@ class Openads_Schema_Manager
      */
     function linkAdd($table_name, $link_add, $link_add_target)
     {
-        $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
-        $links[$table_name][$link_add] = $link_add_target;
-        return Openads_Links::writeLinksDotIni($this->links_trans, $links);
+        if (!empty($this->links_trans)) {
+            $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
+            $links[$table_name][$link_add] = $link_add_target;
+            return Openads_Links::writeLinksDotIni($this->links_trans, $links);
+        }
+        return true;
     }
 
     /**
@@ -654,9 +681,13 @@ class Openads_Schema_Manager
      */
     function readForeignKeys($table_name)
     {
-        $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
-        if (!isset($links[$table_name])) {
-            $links[$table_name] = array();
+        if (!empty($this->links_trans)) {
+            $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
+            if (!isset($links[$table_name])) {
+                $links[$table_name] = array();
+            }
+        } else {
+            $links = array();
         }
         return $links;
     }
@@ -791,32 +822,35 @@ class Openads_Schema_Manager
      */
     function updateFieldLinkRelations($table_name, $field_name_old, $field_name_new)
     {
-        $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
-        reset($links);
-        foreach ($links AS $table => $keys)
-        {
-            foreach ($keys AS $field => $target)
+        if (!empty($this->links_trans)) {
+            $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
+            reset($links);
+            foreach ($links AS $table => $keys)
             {
-                if (($target['table']==$table_name) && ($target['field']==$field_name_old))
+                foreach ($keys AS $field => $target)
                 {
-                    if ($field_name_new)
+                    if (($target['table']==$table_name) && ($target['field']==$field_name_old))
                     {
-                        $target['field'] = $field_name_new;
+                        if ($field_name_new)
+                        {
+                            $target['field'] = $field_name_new;
+                            $links[$table][$field] = $target;
+                        }
+                    }
+                    if (($table==$table_name) && ($field==$field_name_old))
+                    {
+                        if ($field_name_new)
+                        {
+                            unset($links[$table][$field]);
+                            $field = $field_name_new;
+                        }
                         $links[$table][$field] = $target;
                     }
                 }
-                if (($table==$table_name) && ($field==$field_name_old))
-                {
-                    if ($field_name_new)
-                    {
-                        unset($links[$table][$field]);
-                        $field = $field_name_new;
-                    }
-                    $links[$table][$field] = $target;
-                }
             }
+            return Openads_Links::writeLinksDotIni($this->links_trans, $links);
         }
-        return Openads_Links::writeLinksDotIni($this->links_trans, $links);
+        return true;
     }
 
     /**
@@ -828,22 +862,25 @@ class Openads_Schema_Manager
      */
     function deleteTableLinkRelations($table_name)
     {
-        $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
-        foreach ($links AS $table => $keys)
-        {
-            if (($table==$table_name))
+        if (!empty($this->links_trans)) {
+            $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
+            foreach ($links AS $table => $keys)
             {
-                unset($links[$table]);
-            }
-            foreach ($keys AS $field => $target)
-            {
-                if (($target['table']==$table_name))
+                if (($table==$table_name))
                 {
-                    unset($links[$table][$field]);
+                    unset($links[$table]);
+                }
+                foreach ($keys AS $field => $target)
+                {
+                    if (($target['table']==$table_name))
+                    {
+                        unset($links[$table][$field]);
+                    }
                 }
             }
+            return Openads_Links::writeLinksDotIni($this->links_trans, $links);
         }
-        return Openads_Links::writeLinksDotIni($this->links_trans, $links);
+        return true;
     }
 
     /**
@@ -856,24 +893,31 @@ class Openads_Schema_Manager
      */
     function updateTableLinkRelations($table_name, $table_name_new)
     {
-        $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
-        if (isset($links[$table_name]))
-        {
-            $links[$table_name_new] = $links[$table_name];
-            unset($links[$table_name]);
-        }
-
-        foreach ($links AS $table => $keys)
-        {
-            foreach ($keys AS $field => $target)
+        if (!empty($this->links_trans)) {
+            $links = Openads_Links::readLinksDotIni($this->links_trans, $table_name);
+            if (isset($links[$table_name]))
             {
-                if (($target['table']==$table_name))
+                $links[$table_name_new] = $links[$table_name];
+                unset($links[$table_name]);
+            }
+
+            foreach ($links AS $table => $keys)
+            {
+                foreach ($keys AS $field => $target)
+
                 {
-                    $links[$table][$field]['table'] = $table_name_new;
+                    foreach ($keys AS $field => $target)
+                    {
+                        if (($target['table']==$table_name))
+                        {
+                            $links[$table][$field]['table'] = $table_name_new;
+                        }
+                    }
                 }
             }
+            return Openads_Links::writeLinksDotIni($this->links_trans, $links);
         }
-        return Openads_Links::writeLinksDotIni($this->links_trans, $links);
+        return true;
     }
 
     /**
@@ -911,9 +955,11 @@ class Openads_Schema_Manager
             $this->schema_trans,
             $this->links_final,
             $this->links_trans,
+            $this->links_final,
             $this->changes_final,
             $this->changes_trans,
-            MAX_SCHEMA_LOG
+            MAX_SCHEMA_LOG,
+            MAX_PATH.'/www/devel/schema/schema.js'
         );
 
         $aDirs = array(
@@ -922,6 +968,9 @@ class Openads_Schema_Manager
 
         $aErrors = array();
         foreach ($aFiles as $file) {
+            if (empty($file)) {
+                continue;
+            }
             $dir = dirname($file);
             if (!file_exists($file)) {
                 if (!isset($aDirs[$dir])) {
